Save metadata to dictionary and pickle
All the metadata we might want to have access to (e.g. the digit-bounding
boxes, and the labels for the digits that our images contain) is stored
in the included digitStruct.mat file.

Unfortunately, I don't know how to work with HDF5 objects, and a fairly
lengthy search for relevant resources on the interwebs on this topic
has been embarrassingly fruitless.

Accordingly, using a pipeline outlined by `master_yoda` at
https://discussions.udacity.com/t/how-to-deal-with-mat-files/160657/3
we save the image metadata to a dictionary, whose keys are 'height', 'label',
etc.

Then, in accordance with that same pipeline, we pickle the data so that we
don't have to recreate it each time we open the `jupyter` notebook.